androidclient: Added support for float input/output meal/insulin profiles. Fixed some old mapping issues on entities.

+++ androidclient/app/src/main/java/pt/ipl/isel/leic/ps/androidclient/data/api/dto/input/meal/NutritionalInfoInput.kt
@@ -1,7 +1,7 @@
 package pt.ipl.isel.leic.ps.androidclient.data.api.dto.input.meal
 
 class NutritionalInfoInput(
-    val carbs: Int,
-    val amount: Int,
+    val carbs: Float,
+    val amount: Float,
     val unit: String
 )
