+++ androidclient/app/src/main/java/pt/ipl/isel/leic/ps/androidclient/data/db/entity/DbMealInfoEntity.kt
@@ -9,7 +9,7 @@ open class DbMealInfoEntity(
     val submissionId: Int?,
     val restaurantSubmissionId: String?,
     val name: String,
-    val carbs: Int,
+    val carbs: Float,
     val amount: Float,
     val unit: Int,
     val imageUri: String?,
